requirements.txt
asgiref==3.8.1
Django==5.0.6
gunicorn==22.0.0
psycopg2-binary==2.9.11
sqlparse==0.5.0
whitenoise==6.7.0
tweepy==4.16.0
celery==5.5.3
redis==7.0.0
django-celery-beat==2.8.1
djangorestframework==3.16.1
django-filter==24.3
python-dotenv==1.1.1